+++ flake.nix
@@ -2,24 +2,37 @@
   description = "Haskell 'network-simple' libraries";
 
   inputs = {
-    nixpkgs.url =
-      "github:NixOS/nixpkgs?rev=21eda9bc80bef824a037582b1e5a43ba74e92daa";
-    flake-parts.url = "github:hercules-ci/flake-parts";
+    flakety.url = "github:k0001/flakety";
+    nixpkgs.follows = "flakety/nixpkgs";
+    flake-parts.follows = "flakety/flake-parts";
   };
 
   outputs = inputs@{ ... }:
     inputs.flake-parts.lib.mkFlake { inherit inputs; } {
-      flake.overlays.default = final: prev: {
-        haskell = prev.haskell // {
-          packageOverrides = prev.lib.composeExtensions
-            (prev.haskell.packageOverrides or (_: _: { })) (hself: hsuper: {
-              network-simple = hself.callPackage ./network-simple { };
-              network-simple-tls = hself.callPackage ./network-simple-tls { };
-              network-simple-ws = hself.callPackage ./network-simple-ws { };
-              network-simple-wss = hself.callPackage ./network-simple-wss { };
-            });
-        };
-      };
+      flake.overlays.default = inputs.nixpkgs.lib.composeManyExtensions [
+        inputs.flakety.overlays.default
+        (final: prev:
+          let
+            hsLib = prev.haskell.lib;
+            hsClean = drv:
+              hsLib.overrideCabal drv
+              (old: { src = prev.lib.sources.cleanSource old.src; });
+          in {
+            haskell = prev.haskell // {
+              packageOverrides = prev.lib.composeExtensions
+                (prev.haskell.packageOverrides or (_: _: { })) (hself: hsuper: {
+                  network-simple =
+                    hsClean (hself.callPackage ./network-simple { });
+                  network-simple-tls =
+                    hsClean (hself.callPackage ./network-simple-tls { });
+                  network-simple-ws =
+                    hsClean (hself.callPackage ./network-simple-ws { });
+                  network-simple-wss =
+                    hsClean (hself.callPackage ./network-simple-wss { });
+                });
+            };
+          })
+      ];
       systems = [ "x86_64-linux" "i686-linux" "aarch64-linux" ];
       perSystem = { config, system, pkgs, ... }: {
         _module.args.pkgs = import inputs.nixpkgs {
@@ -30,54 +43,47 @@
           default = pkgs.releaseTools.aggregate {
             name = "every output from this flake";
             constituents = [
-              config.packages.network-simple__ghc8107
-              config.packages.network-simple__ghc8107.doc
-              config.packages.network-simple__ghc925
-              config.packages.network-simple__ghc925.doc
-              config.packages.network-simple__ghc943
-              config.packages.network-simple__ghc943.doc
+              config.packages.network-simple__ghc98
+              config.packages.network-simple__ghc98.doc
+              config.packages.network-simple__ghc96
+              config.packages.network-simple__ghc96.doc
 
-              config.packages.network-simple-tls__ghc8107
-              config.packages.network-simple-tls__ghc8107.doc
-              config.packages.network-simple-tls__ghc925
-              config.packages.network-simple-tls__ghc925.doc
-              config.packages.network-simple-tls__ghc943
-              config.packages.network-simple-tls__ghc943.doc
+              config.packages.network-simple-tls__ghc98
+              config.packages.network-simple-tls__ghc98.doc
+              config.packages.network-simple-tls__ghc96
+              config.packages.network-simple-tls__ghc96.doc
 
-              config.packages.network-simple-ws__ghc8107
-              config.packages.network-simple-ws__ghc8107.doc
-              config.packages.network-simple-ws__ghc925
-              config.packages.network-simple-ws__ghc925.doc
-              config.packages.network-simple-ws__ghc943
-              config.packages.network-simple-ws__ghc943.doc
+              config.packages.network-simple-ws__ghc98
+              config.packages.network-simple-ws__ghc98.doc
+              config.packages.network-simple-ws__ghc96
+              config.packages.network-simple-ws__ghc96.doc
 
-              config.packages.network-simple-wss__ghc8107
-              config.packages.network-simple-wss__ghc8107.doc
-              config.packages.network-simple-wss__ghc925
-              config.packages.network-simple-wss__ghc925.doc
-              config.packages.network-simple-wss__ghc943
-              config.packages.network-simple-wss__ghc943.doc
+              config.packages.network-simple-wss__ghc98
+              config.packages.network-simple-wss__ghc98.doc
+              config.packages.network-simple-wss__ghc96
+              config.packages.network-simple-wss__ghc96.doc
 
-              config.devShells.ghc8107
-              config.devShells.ghc925
-              config.devShells.ghc943
+              config.devShells.ghc98
+              config.devShells.ghc96
             ];
           };
-          network-simple__ghc8107 = pkgs.haskell.packages.ghc8107.network-simple;
-          network-simple__ghc925 = pkgs.haskell.packages.ghc925.network-simple;
-          network-simple__ghc943 = pkgs.haskell.packages.ghc943.network-simple;
+          network-simple__ghc98 = pkgs.haskell.packages.ghc98.network-simple;
+          network-simple__ghc96 = pkgs.haskell.packages.ghc96.network-simple;
 
-          network-simple-tls__ghc8107 = pkgs.haskell.packages.ghc8107.network-simple-tls;
-          network-simple-tls__ghc925 = pkgs.haskell.packages.ghc925.network-simple-tls;
-          network-simple-tls__ghc943 = pkgs.haskell.packages.ghc943.network-simple-tls;
+          network-simple-tls__ghc98 =
+            pkgs.haskell.packages.ghc98.network-simple-tls;
+          network-simple-tls__ghc96 =
+            pkgs.haskell.packages.ghc96.network-simple-tls;
 
-          network-simple-ws__ghc8107 = pkgs.haskell.packages.ghc8107.network-simple-ws;
-          network-simple-ws__ghc925 = pkgs.haskell.packages.ghc925.network-simple-ws;
-          network-simple-ws__ghc943 = pkgs.haskell.packages.ghc943.network-simple-ws;
+          network-simple-ws__ghc98 =
+            pkgs.haskell.packages.ghc98.network-simple-ws;
+          network-simple-ws__ghc96 =
+            pkgs.haskell.packages.ghc96.network-simple-ws;
 
-          network-simple-wss__ghc8107 = pkgs.haskell.packages.ghc8107.network-simple-wss;
-          network-simple-wss__ghc925 = pkgs.haskell.packages.ghc925.network-simple-wss;
-          network-simple-wss__ghc943 = pkgs.haskell.packages.ghc943.network-simple-wss;
+          network-simple-wss__ghc98 =
+            pkgs.haskell.packages.ghc98.network-simple-wss;
+          network-simple-wss__ghc96 =
+            pkgs.haskell.packages.ghc96.network-simple-wss;
         };
         devShells = let
           shellFor = hpkgs:
@@ -92,10 +98,9 @@
               nativeBuildInputs = [ pkgs.cabal-install pkgs.cabal2nix ];
             };
         in {
-          default = config.devShells.ghc943;
-          ghc8107 = shellFor pkgs.haskell.packages.ghc8107;
-          ghc925 = shellFor pkgs.haskell.packages.ghc925;
-          ghc943 = shellFor pkgs.haskell.packages.ghc943;
+          default = config.devShells.ghc98;
+          ghc98 = shellFor pkgs.haskell.packages.ghc98;
+          ghc96 = shellFor pkgs.haskell.packages.ghc96;
         };
       };
     };