+++ network-simple-tls/default.nix
@@ -1,15 +1,16 @@
-{ mkDerivation, base, bytestring, data-default, lib, network
-, network-simple, safe-exceptions, tls, tls-session-manager
-, transformers, x509, x509-store, x509-system, x509-validation
+{ mkDerivation, base, bytestring, crypton-x509, crypton-x509-store
+, crypton-x509-system, crypton-x509-validation, data-default, lib
+, network, network-simple, safe-exceptions, tls
+, tls-session-manager, transformers
 }:
 mkDerivation {
   pname = "network-simple-tls";
   version = "0.4.1";
-  src = lib.sources.cleanSource ./.;
+  src = ./.;
   libraryHaskellDepends = [
-    base bytestring data-default network network-simple safe-exceptions
-    tls tls-session-manager transformers x509 x509-store x509-system
-    x509-validation
+    base bytestring crypton-x509 crypton-x509-store crypton-x509-system
+    crypton-x509-validation data-default network network-simple
+    safe-exceptions tls tls-session-manager transformers
   ];
   homepage = "https://hackage.haskell.org/package/network-simple-tls";
   description = "Simple interface to TLS secured network sockets";