+++ network-simple/default.nix
@@ -4,7 +4,7 @@
 mkDerivation {
   pname = "network-simple";
   version = "0.4.5";
-  src = lib.sources.cleanSource ./.;
+  src = ./.;
   libraryHaskellDepends = [
     base bytestring network network-bsd safe-exceptions socks
     transformers